Duplicate Code Reduction
Implemented method to reduce duplicate code. Now when code such as the event summary and event content code is duplicated on multiple pages you can create a php file with the duplicate code and place it in the template-parts folder and call it instead of having the code in the pages.