Added professor like system
Added the ability for users to "like" professors.

Logged in users will be able to "like" a professor as well as remove their "like" from a professors profile.

People that are not logged in can see the like count but they cannot add or remove a like from the professor.